Refactor texture scaling vector construction

+++ BasicShmup/Src/Entities/Projectiles/IProjectileConfiguration.cs
@@ -8,5 +8,5 @@ public interface IProjectileConfiguration
     public Speed Speed { get; }
     public float ColliderRadius { get; }
     public Texture2D Texture { get; }
-    public int TextureScaling { get; }
+    public float TextureScaling { get; }
 }
